Created a typst generator
